MOSYNC-3096, MOSYNC-3097: added a check inside MoSyncUtil RunActionOnMainThread that checks if we're on the current thread and runs the action; Moved most of the business logic for handling syscalls asyncronously inside MoSyncNativeUIAsync.cs; large refactorization of the current mosync runtime - added many regions(for a better lisibility) and grouped methods based on their scope
