+++ runtimes/csharp/windowsphone/mosync/mosyncRuntime/Source/Modules/NativeUI/MoSyncVerticalLayout.cs
@@ -99,7 +99,6 @@ namespace MoSync
                 mGrid.ColumnDefinitions.Add(mSpacerRight);
 
                 View = mGrid;
-//                mView = mGrid;
 
                 mGrid.Margin = new Thickness(0.0);
 
@@ -119,8 +118,8 @@ namespace MoSync
             public override void AddChild(IWidget child)
             {
                 base.AddChild(child);
-     //           MoSync.Util.RunActionOnMainThreadSync(() =>
-     //           {
+                MoSync.Util.RunActionOnMainThreadSync(() =>
+                {
                     // The logic behind this layout system is to create a stack panel for each
                     // MoSync widget. The stack panel is used for wrapping and filling size policies.
                     // When the widget size policy states that it should fill on both horizontal and
@@ -224,7 +223,7 @@ namespace MoSync
 
                     widget.RowNumber = mGrid.RowDefinitions.Count - 2;
                     widget.ColumnNumber = 1;
-           //     });
+                });
             }
 
             /**
