+++ runtimes/csharp/windowsphone/mosync/mosyncRuntime/mosyncRuntime.csproj
@@ -103,6 +103,7 @@
     <Compile Include="Source\Modules\MoSyncScreenOrientationModule.cs" />
     <Compile Include="Source\Modules\MoSyncSensorsModule.cs" />
     <Compile Include="Source\Modules\MoSyncSoundModule.cs" />
+    <Compile Include="Source\Modules\NativeUI\MoSyncNativeUIAsync.cs" />
     <Compile Include="Source\Modules\NativeUI\MoSyncScreenTransitions.cs" />
     <Compile Include="Source\Modules\NativeUI\ListView\ListViewModel\ListItem.cs" />
     <Compile Include="Source\Modules\NativeUI\ListView\ListViewModel\ListSection.cs" />
